[Refactor] 🍎 길찾기 페이지 코드 리팩토링
