+++ src/components/Direction/BottomFloatingBox/BankInfo.tsx
