+++ src/components/Direction/TopSheet/TotalTime.tsx
@@ -1,0 +1,16 @@
+import { totalTimeAtom } from '@/stores';
+import { useAtom } from 'jotai';
+
+export default function TotalTime() {
+  const [totalTime] = useAtom(totalTimeAtom);
+
+  return (
+    <div className='flex items-end gap-3'>
+      <div>소요시간</div>
+      <div className='flex items-end'>
+        <div className='text-3xl font-bold'>{(totalTime / 60).toFixed(0)}</div>
+        <div>분</div>
+      </div>
+    </div>
+  );
+}
